rename: 월 탐색 컴포넌트 추가 및 Home 페이지에서 임포트 수정
