+++ src/pages/Home.tsx
@@ -1,6 +1,6 @@
 import CreateExpense from "../components/CreateExpense";
 import ExpenseList from "../components/ExpenseList";
-import MouthNavigation from "../components/MouthNavigation";
+import MouthNavigation from "../components/MonthNavigation";
 
 const Home = () => {
   return (
